2022年7月14日
ad14n_ad15n-release_v1.5.0
SDK等级： B
发布时间：2022年7月14日
发布版本：ad14n-ad15n-release_V1.5.0
发布说明：
以下改动对应”ad140-release_v1.2.1” 和“ad150-release_v1.2.1”；
1、修复A0芯片外挂部分flash（5e3213/5e3214）运行一线模式powerdown无法唤醒问题；
2、修复mio模块pwm公式错误导致满占空比时会有脉冲问题；
3、修复带电池升级中途断电，无法重新升级问题；
4、修复midi主副旋一起OKON无法出声问题；
5、修复midi小节回退功能异常问题；
6、修复USB DM/DP作唤醒脚时，内部上下拉使能pullup_down_enable无法关闭问题；
7、优化powerdown唤醒DAC初始化时间，由20ms缩短至1.2ms；
8、新增flash安全寄存器驱动；
9、新增ini自定义空间作为虚拟设备管理驱动；
以下改动对应”ad140-release_v1.3.0” 和“ad150-release_v1.3.0”；
10、修复大容量flash（>16M字节）下载时出现CRC校验错误问题；
11、修复大容量flash（>16M字节）powerdown唤醒可能死机的问题；
12、修复解码断点检查可能出现越界访问的问题；
13、修复tab_read读表模块初始化无效的问题；
同步gitee修复：FAQ：tab_read读表模块bug修复
14、修复开机解码可能出现po声问题，在电压扫描初始化时切换时钟；
同步gitee修复：FAQ：AD14N / AC104N上电解码可能有po声？
15、更新Voice_Pitch变音算法库，新增卡通音效，并完善其他音效的配置；
16、更新Notch_Howling陷波啸叫抑制算法库，使用新的啸叫抑制算法；
17、更新Howling_PitchShifter移频啸叫抑制算法库，使用新的啸叫抑制算法；
18、更新MIDI库，支持使用压缩的mda音色库文件，
（mda音色库需使用新的midi工具制作，运行压缩库时系统时钟需要至少96MHz）

19、优化midi解码只有主旋OKON副旋不发声时的响应速度；
20、新增midi解码只有主旋OKON副旋不发声时，可使用快速打断模式；
21、新增midi琴应用，并使用新的mda音色库，应用内容包括；
a)琴键的按下与抬起响应；
b)乐器通道切换；
c)颤音；
d)弯音轮；
e)音符开始与音符结束回调；
f)叠加A解码；
22、新增低电检测方式，可配置使用lvd电压检测或使用固定值检测；
23、新增OTA升级时I/O保持功能；
同步gitee改动：FAQ：AD14N / AC104N V1.2.1 SDK如何在OTA设备升级过程中保持IO状态？
24、优化powerdown流程，增强稳定性；
25、优化下载目录生产烧录文件对齐方式，强制只生成一份4K字节对齐的flash.bin；
同步gitee改动：FAQ：关于FLASH数据工程文件对齐方式对升级下载的影响
26、优化MIC_RES_BUILT_IN配置，现可配置mic使用外部偏置；
27、优化VM擦除方式，加快256字节对齐flash的VM擦除速度；
28、优化lvd接口，触发lvd后可配置直接复位或进入lvd中断响应函数；
29、优化vfs管理，将fat独有的函数移出vfs.c文件；
30、优化msg驱动，获取事件消息关中断时AUDIO和TICKTIMER中断不会被打断；
31、优化norflash驱动，精简结构体大小，读写0字节直接返回；
32、优化mio启动判断，现可支持不同的DAC采样率；
33、优化flash安全寄存器驱动，可支持flash单线模式；
34、优化LD文件部分段名以及段位置，函数放ram的段修改为统一格式(.*.text.cache.L1)；
35、优化下载目录结构，app_mg工程和simple工程拆分独立的下载目录；
36、优化ticktimer_loop架构，拆分成sys、modules和app来执行定时函数；
37、优化F1A解码库，提高解码效率；
（LD文件中(*.f1a_code)段位置不可移动或调整，否则解码可能不正常！）
38、优化vm擦写时不受关中断影响的中断数，由一个改为任意多个；
（放出的中断函数及其调用的函数/变量必须放在ram里，否则会导致死机！）
39、完善SDK说明手册，部分章节添加示例说明；

文档更新：
1、新增文档《杰理AD14N_AD15N_MIDI应用说明文档.pdf》；
2、新增文档《杰理AD14N_AC104N音效算法应用介绍文档.pdf》；
3、更新了芯片规格书《AD156A_Datasheet V1.2.pdf》，增加了LRC温差和时钟精度说明；

工具版本：
1、一拖二烧写器：V2.26.6，通过软件上远程更新；
2、一拖八烧写器：V3.1.11，通过软件上远程更新；
3、音频文件转换工具：V1.2.17，在gitee上下载；
4、JLmidi_new_0706工具（新增）：该工具只能用于此发布版本以及之后的SDK，在gitee上下载，gitee上会保留旧工具（JLmidi_60）；
5、pcm_eq工具（新增）：V3.1.7，在gitee上下载；

+++ sdk/app/bsp/common/norflash/norflash.h
@@ -29,8 +29,8 @@
 // static inline void os_mutex_post(OS_MUTEX *mutex) {}
 #include "jiffies.h"
 #include "wdt.h"
-typedef volatile u32 flash_mutex;
-static inline void flash_mutex_create(flash_mutex *sem, u32 count)//初始化
+typedef volatile u8 flash_mutex;
+static inline void flash_mutex_create(flash_mutex *sem, u8 count)//初始化
 {
     *sem = count;
 }
@@ -38,7 +38,7 @@ static inline void flash_mutex_post(flash_mutex *sem)//
 {
     (*sem) = 1;
 }
-static inline s32 flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为timeout=0时，死等
+static inline s8 flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为timeout=0时，死等
 {
     u32 _timeout = timeout + jiffies;
     extern void wdt_clear();
@@ -54,7 +54,7 @@ static inline s32 flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为timeou
     }
     return 0;
 }
-static inline s32 query_flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为timeout=0时，死等
+static inline s8 query_flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为timeout=0时，死等
 {
     while (1) {
         if (*sem) {
@@ -65,7 +65,7 @@ static inline s32 query_flash_mutex_pend(flash_mutex *sem, u32 timeout)// 当为
     }
     return 0;
 }
-static inline void flash_mutex_set(flash_mutex *sem, u32 count)
+static inline void flash_mutex_set(flash_mutex *sem, u8 count)
 {
     *sem = count;
 }
@@ -102,9 +102,9 @@ enum {
 
 
 struct norflash_dev_platform_data {
-    int spi_hw_num;         //只支持SPI1或SPI2
-    u32 spi_cs_port;        //cs的引脚
-    u32 spi_read_width;     //flash读数据的线宽
+    s8 spi_hw_num;         //只支持SPI1或SPI2
+    u8 spi_cs_port;        //cs的引脚
+    u8 spi_read_width;     //flash读数据的线宽
     const struct spi_platform_data *spi_pdata;
     u32 start_addr;         //分区起始地址
     u32 size;               //分区大小，若只有1个分区，则这个参数可以忽略
